requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pydantic[email]
email-validator
pyjwt
bcrypt
pyjwt
smtplib
email-validator